+++ inst/rmarkdown/templates/rnaseq/skeleton/DE/load_data.R
@@ -1,6 +1,6 @@
 library(tidyverse)
 library(SummarizedExperiment)
-
+library(janitor)
 load_metrics <- function(se_object, multiqc_data_dir){
 
   # bcbio input
@@ -8,7 +8,7 @@ load_metrics <- function(se_object, multiqc_data_dir){
 
     se <- readRDS(se_object)
     metrics <- metadata(se)$metrics %>% mutate(sample = toupper(sample)) %>% as.data.frame()
-      # left_join(coldata %>% rownames_to_column('sample')) %>% column_to_rownames('sample')
+    # left_join(coldata %>% rownames_to_column('sample')) %>% column_to_rownames('sample')
 
   } else { #nf-core input
 
@@ -28,15 +28,20 @@ load_metrics <- function(se_object, multiqc_data_dir){
       group_by(sample) %>%
       summarize(total_reads = sum(fastqc_raw_total_sequences))
 
+    if (!("custom_content_biotype_counts_percent_r_rna" %in% colnames(metrics))){
+      metrics[["custom_content_biotype_counts_percent_r_rna"]] <- rep(0,nrow(metrics))
+    }
+    # browser()
+
     metrics <- metrics %>%
-      filter(!grepl('_', sample)) %>%
+      filter(!grepl('_[12]$', sample)) %>%
       remove_empty(which = 'cols') %>%
       full_join(total_reads) %>%
-      mutate(mapped_reads = samtools_reads_mapped) %>%
+      dplyr::rename(mapped_reads = samtools_reads_mapped) %>%
       mutate(exonic_rate = exonic/(star_uniquely_mapped * 2)) %>%
       mutate(intronic_rate = intronic/(star_uniquely_mapped * 2)) %>%
       mutate(intergenic_rate = intergenic/(star_uniquely_mapped * 2)) %>%
-      mutate(r_rna_rate = custom_content_biotype_counts_percent_r_rna) %>%
+      dplyr::rename(r_rna_rate = custom_content_biotype_counts_percent_r_rna) %>%
       mutate(x5_3_bias = qualimap_5_3_bias) %>%
       dplyr::select(
         sample,
@@ -47,17 +52,26 @@ load_metrics <- function(se_object, multiqc_data_dir){
         intergenic_rate,
         r_rna_rate,
         x5_3_bias
-      ) %>% as.data.frame()
+      ) %>%
+      dplyr::select(where(~!all(is.na(.)))) %>%
+      as.data.frame()
 
     # metrics <- metrics %>%
     #   full_join(meta_df , by = c("sample" = "sample"))
   }
+  metrics$sample <- make.names(metrics$sample)
   rownames(metrics) <- metrics$sample
   return(metrics)
 }
 
 load_coldata <- function(coldata_fn, column, numerator, denominator, subset_column = NULL, subset_value = NULL){
-  coldata=read.csv(coldata_fn)
+  coldata=read.csv(coldata_fn) %>%
+    dplyr::select(!matches("fastq") & !matches("strandness")) %>%
+    distinct()
+  if('description' %in% names(coldata)){
+    coldata$sample <- coldata$description
+  }
+  coldata <- coldata %>% distinct(sample, .keep_all = T)
   stopifnot(column %in% names(coldata))
 
   # use only some samples, by default use all
@@ -65,14 +79,11 @@ load_coldata <- function(coldata_fn, column, numerator, denominator, subset_colu
     coldata <- coldata[coldata[[paste(subset_column)]] == subset_value, ]
   }
   #coldata <- coldata[coldata[[paste(column)]] %in% c(numerator, denominator), ]
-
-  if('description' %in% names(coldata)){
-    rownames(coldata) <- coldata$description
-  }else {
-    rownames(coldata) <- coldata$sample
-    coldata$description <- coldata$sample
-  }
   #browser()
+  coldata$sample <- make.names(coldata$sample)
+  rownames(coldata) <- coldata$sample
+  coldata$description <- coldata$sample
+
   coldata[[column]] = relevel(as.factor(coldata[[column]]), denominator)
 
   return(coldata)
@@ -85,7 +96,7 @@ load_counts <- function(counts_fn){
     counts <- read_csv(counts_fn) %>% column_to_rownames('gene')
   } else { # nf-core input
     counts <- read_tsv(counts_fn) %>% dplyr::select(-gene_name) %>%
-      mutate(gene_id = str_sub(gene_id, 1, 15)) %>%
+      mutate(gene_id = str_replace(gene_id, pattern = "\\.[0-9]$", "")) %>%
       column_to_rownames('gene_id') %>% round
 
     return(counts)