+++ NEWS.md
@@ -1,6 +1,11 @@
 # bcbioR 0.1.2
 
 * Adapt templates to nf-core rnaseq
+* Fix when sample start by number
+* Fix when rRNA rate is missing
+* Add by sample plots in QC
+* Add function to check nfcore samplesheet
+* Add PCA with variance analysis
 
 # bcbioR 0.1.1